requirements.txt
# Complete Dependencies for Forex Trading RL Agent
# ==============================================
# Based on currently installed and tested packages

# Core Python Infrastructure
certifi>=2025.10.5
charset-normalizer>=3.4.4
idna>=3.11
requests>=2.32.5
urllib3>=2.5.0
setuptools>=80.9.0

# Core Scientific Computing
numpy>=2.2.6
pandas>=2.3.3
scipy>=1.16.3
scikit-learn>=1.8.0

# Technical Analysis & Numerical Computing
pandas-ta>=0.4.71b0
numba>=0.61.2
llvmlite>=0.44.0

# Reinforcement Learning
stable-baselines3>=2.7.1
gymnasium>=1.2.2
torch>=2.9.1
torchvision>=0.24.1
torchaudio>=2.9.1

# Hyperparameter Optimization
optuna>=4.6.0

# Visualization and Plotting
matplotlib>=3.10.8
seaborn>=0.13.2
pillow>=12.0.0
contourpy>=1.3.3
cycler>=0.12.1
fonttools>=4.61.1
kiwisolver>=1.4.9
pyparsing>=3.2.5

# Logging and Monitoring
tensorboard>=2.20.0
tensorboard-data-server>=0.7.2
tqdm>=4.67.1
colorlog>=6.10.1

# Configuration and Data
pyyaml>=6.0.3
joblib>=1.5.2
cloudpickle>=3.1.2

# Database and Migration
sqlalchemy>=2.0.45
alembic>=1.17.2
greenlet>=3.3.0
mako>=1.3.10

# Utilities and Support
psutil>=7.1.3
packaging>=25.0
typing-extensions>=4.15.0
filelock>=3.20.0
fsspec>=2025.12.0
networkx>=3.6.1
sympy>=1.14.0
mpmath>=1.3.0

# Date and Time
python-dateutil>=2.9.0
pytz>=2025.2
tzdata>=2025.3

# Web and Communication
grpcio>=1.76.0
jinja2>=3.1.6
markdown>=3.10
markupsafe>=3.0.3
werkzeug>=3.1.4
absl-py>=2.3.1
protobuf>=6.33.2

# Gymnasium Dependencies
farama-notifications>=0.0.4

# Additional Utilities
colorama>=0.4.6
six>=1.17.0
threadpoolctl>=3.6.0
brotli>=1.1.0

# Optional Development Tools
altgraph>=0.17.4
mutagen>=1.47.0
pefile>=2023.2.7
pycryptodomex>=3.23.0
pyinstaller>=6.16.0
pyinstaller-hooks-contrib>=2025.9
pywin32-ctypes>=0.2.3
websockets>=15.0.1
yt-dlp>=2025.10.22

# Statistical Analysis (Optional)
# statsmodels>=0.14.0

# Development Tools (Optional)
# jupyter>=1.0.0
# ipykernel>=6.25.0
# black>=23.0.0
# flake8>=6.0.0